Handle "tus: upload not found" (scenario 2)
## Issue
I removed the hardcoded delay between "tus uploaded" and "notify" to see if it was still required, since we'll be doing QA. While I couldn't hit the issue, Miko still did (location dependent?).

I'm guessing the hardcoded delay is not sufficient for all cases, and thus explains the remaining "tus: upload not found" occurrences.

## Change
Improved the handling by checking the response of `notify` and retrying (instead of using hardcoded delay).